requirements-dev.txt
# Development dependencies
pytest>=8.0.0
pytest-httpx>=0.30.0
ruff>=0.3.0
mypy>=1.9.0
